+++ neoforge/src/main/java/de/maxhenkel/voicechat/intercompatibility/NeoForgeClientCompatibilityManager.java
@@ -1,15 +1,11 @@
 package de.maxhenkel.voicechat.intercompatibility;
 
 import com.mojang.blaze3d.platform.InputConstants;
-import com.mojang.blaze3d.vertex.PoseStack;
 import de.maxhenkel.voicechat.events.ClientVoiceChatConnectedEvent;
 import de.maxhenkel.voicechat.events.ClientVoiceChatDisconnectedEvent;
 import de.maxhenkel.voicechat.voice.client.ClientVoicechatConnection;
 import net.minecraft.client.KeyMapping;
 import net.minecraft.client.Minecraft;
-import net.minecraft.client.renderer.SubmitNodeCollector;
-import net.minecraft.client.renderer.entity.state.EntityRenderState;
-import net.minecraft.client.renderer.state.CameraRenderState;
 import net.minecraft.client.server.IntegratedServer;
 import net.minecraft.network.Connection;
 import net.minecraft.server.packs.repository.RepositorySource;
@@ -58,14 +54,9 @@ public class NeoForgeClientCompatibilityManager extends ClientCompatibilityManag
         keyMappings = new CopyOnWriteArrayList<>();
     }
 
-    //TODO Implement once NeoForge fixes this
-    /*@SubscribeEvent
+    @SubscribeEvent
     public void onRenderName(net.neoforged.neoforge.client.event.RenderNameTagEvent.DoRender event) {
-        renderNameplateEvents.forEach(renderNameplateEvent -> renderNameplateEvent.render(event.getEntityRenderState(), null, event.getPoseStack(), event.getSubmitNodeCollector()));
-    }*/
-
-    public void onRenderNameplate(EntityRenderState renderState, CameraRenderState cameraRenderState, PoseStack stack, SubmitNodeCollector collector) {
-        renderNameplateEvents.forEach(renderNameplateEvent -> renderNameplateEvent.render(renderState, cameraRenderState, stack, collector));
+        renderNameplateEvents.forEach(renderNameplateEvent -> renderNameplateEvent.render(event.getEntityRenderState(), event.getCameraRenderState(), event.getPoseStack(), event.getSubmitNodeCollector()));
     }
 
     @SubscribeEvent