+++ neoforge/changelog.md
@@ -1,1 +1,1 @@
-- Updated to 1.21.9
+- Fixed NeoForge name render event not being used for icons